+++ index.tsx
@@ -1,1 +1,2 @@
 export { FolderStructure } from "./src/FolderStructure";
+export { SkipLink } from "./src/SkipLink";